add landing component
