+++ src/components/Landing.js
@@ -1,0 +1,18 @@
+import HeaderLanding from './HeaderLanding';
+import MainLanding from './MainLanding';
+
+const Landing = () => {
+  return (
+    <div className="bck-landing">
+      {/* HEADER */}
+      <HeaderLanding></HeaderLanding>
+      {/* HEADER */}
+      {/* MAIN */}
+      <MainLanding></MainLanding>
+    </div>
+
+    /* MAIN  */
+  );
+};
+
+export default Landing;
