+++ src/styles/App.scss
@@ -2,57 +2,34 @@
 @import "core/variables";
 @import "core/functions";
 @import "core/_mixin.scss";
-@import "./layout/MainIndex";
-
-/*@import 'components/code';
-@import 'components/links';*/
-//@import "layout/header";
-//@import "layout/footer";
-/*@import 'pages/index';
-*/
 
 * {
   box-sizing: border-box;
   font-family: $font-Open;
 }
 
-.body {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  background-image: url('../images/background-main-copy.png');
-  background-repeat: no-repeat;
-  background-size: cover;
-}
-
-.backimage {
-  background-image: url('../images/background-main-copy.png');
-  background-repeat: no-repeat;
-  background-size: cover;
-}
-
-.mainCreate {
-  background-image: linear-gradient(to bottom,
-      #fafafa,
-      #cacaca 7%,
-      #fafafa 97%,
-    );
+// .mainCreate {
+//   background-image: linear-gradient(to bottom,
+//       #fafafa,
+//       #cacaca 7%,
+//       #fafafa 97%,
+//     );
 
 
 
-  @media all and (min-width: 768px) {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    min-height: 80vh;
+//   @media all and (min-width: 768px) {
+//     display: grid;
+//     grid-template-columns: 1fr 1fr;
+//     min-height: 80vh;
 
-    &__card {
-      order: 1;
-      margin: 20px 40px 21px 0;
-      border-radius: 6px;
-    }
-  }
+//     &__card {
+//       order: 1;
+//       margin: 20px 40px 21px 0;
+//       border-radius: 6px;
+//     }
+//   }
 
-  @media all and (min-width: 1200px) {
-    grid-template-columns: 1fr 2fr;
-  }
-}
+//   @media all and (min-width: 1200px) {
+//     grid-template-columns: 1fr 2fr;
+//   }
+// }
