+++ src/styles/components/HeaderLanding.scss
@@ -1,3 +1,6 @@
+@import "../core/variables";
+@import "../core/functions";
+@import "../core/_mixin.scss";
 .header {
   display: flex;
   justify-content: center;
